Add HomeController for root endpoint and remove home endpoint from HouseController
